Change: Refactor sub-package 'errors' to 'exceptions' for more clear sub-package structure.
