+++ llyra/exceptions/backends/__init__.py
@@ -1,0 +1,1 @@
+from .remotes import RemoteError
