+++ llyra/exceptions/components/strategys.py
